HW-3 리팩토링 : Movie Adapter에서 MovieViewHolder 분리
 - 코드 양이 많아짐에 따라 분리하여 관리



- 기능(feat): 기능 추가, 삭제, 변경 (제품 코드 수정 발생)
- 버그(fix): 버그 수정 (제품 코드 수정 발생)
- 리팩토링(refactor): 코드 리팩토링 (제품 코드 수정 발생)
- 형식(style): 코드 형식, 정렬 등의 변경 (제품 코드 수정 발생. 하지만 동작에 영향을 주는 변경은 없음)
- 테스트(test): 테스트 코드 추가, 삭제, 변경 등 (제품 코드 수정 없음. 테스트 코드에 관련된 모든 변경에 해당)
- 문서(docs): 문서 추가, 삭제, 변경 (코드 수정 없음)
- 기타: 위에 해당되지 않는 모든 변경(예: 빌드 스크립트 수정, 패키지 배포 설정 변경 등)을 포함 (코드 수정 없음)
