+++ 2005/chuuuul/ArchitectureStudy/app/src/main/java/com/example/architecture/activity/search/adapter/MovieAdapter.kt
@@ -4,24 +4,19 @@ import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
 import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
 import com.example.architecture.R
 import com.example.architecture.data.model.MovieModel
-import com.example.architecture.util.ConstValue.Companion.NO_IMAGE_URL
-import kotlinx.android.synthetic.main.item_movie.view.*
 
-class MovieAdapter :  RecyclerView.Adapter<MovieViewHolder>() {
+class MovieAdapter : RecyclerView.Adapter<MovieHolder>(), MovieAdapterContract.View {
 
     private val movieList = mutableListOf<MovieModel>()
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MovieViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MovieHolder {
 
         val view = LayoutInflater.from(parent.context).inflate(R.layout.item_movie, parent, false)
-        val viewHolder = MovieViewHolder(view)
+        val viewHolder = MovieHolder(view)
 
         view.setOnClickListener {
             openWebPage(it.context, movieList[viewHolder.layoutPosition].link)
@@ -39,10 +34,10 @@ class MovieAdapter :  RecyclerView.Adapter<MovieViewHolder>() {
         return movieList.count()
     }
 
-    override fun onBindViewHolder(holder: MovieViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: MovieHolder, position: Int) {
 
         val movie = movieList[position]
-        holder.onBind( movie )
+        holder.onBind(movie)
     }
 
     fun addNewItems(movieList: List<MovieModel>) {
@@ -55,32 +50,4 @@ class MovieAdapter :  RecyclerView.Adapter<MovieViewHolder>() {
     }
 }
 
-class MovieViewHolder(view: View) : RecyclerView.ViewHolder(view)
-{
-    fun onBind(movie : MovieModel)
-    {
-        itemView.tv_movie_title.text = removeMarkupTag(movie.title)
-        itemView.tv_movie_pubDate.text = movie.pubDate
 
-        itemView.ratingBar_movie_rating.rating = movie.userRating
-
-        setMovieImage(itemView.img_movie_Image, movie.image)
-
-    }
-
-    private fun removeMarkupTag(html: String): String {
-        return html.replace("<b>", "").replace("</b>", "")
-    }
-
-    private fun setMovieImage(imageView: ImageView, imageUrl: String) {
-        val url = if (imageUrl.isBlank()) {
-            NO_IMAGE_URL
-        } else {
-            imageUrl
-        }
-        Glide.with(imageView.context)
-            .load(url).placeholder(R.drawable.ic_loading_black_24dp)
-            .error(R.drawable.image_loaderror).centerCrop().into(imageView)
-    }
-
-}
